backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
openai==1.3.7
pydantic==2.5.2
python-dotenv==1.0.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart
python-jose[cryptography]
pydantic-settings
python-dotenv